Make abstract TestBench class usable from the demo project
