+++ hummingbird-test-util/pom.xml
@@ -11,7 +11,6 @@
 	<name>Testing utilities for Hummingbird</name>
 	<packaging>jar</packaging>
 	<properties>
-		<maven.deploy.skip>true</maven.deploy.skip>
 	</properties>	
 	<dependencies>
 		<dependency>
